+++ shared/src/commonMain/kotlin/org/example/project/data/api/catalog/CatalogApi.kt
@@ -8,6 +8,6 @@ import org.example.project.data.api.catalog.model.GetProductsResponseBody
 interface CatalogApi {
     suspend fun getCategories(): GetCatalogResponseBody
     suspend fun getCategory(id: Long) : GetCategoryResponseBody
-    suspend fun getProduct(productId: Int): GetProductResponseBody
+    suspend fun getProduct(productId: Long): GetProductResponseBody
     suspend fun getProductsByCategory(categoryId: Long): GetProductsResponseBody
 }
