+++ shared/src/commonMain/kotlin/org/example/project/features/catalog/CatalogCallbacks.kt
@@ -3,6 +3,6 @@ package org.example.project.features.catalog
 data class CatalogCallbacks(
     val onBack: () -> Unit,
     val onNavigateToCart: () -> Unit,
-    val showProductCard: (Int) -> Unit,
+    val showProductCard: (Long) -> Unit,
     val onNavigateToCategory: (Int) -> Unit,
 )
