+++ shared/src/commonMain/kotlin/org/example/project/navigation/DialogConfig.kt
@@ -4,7 +4,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 sealed class DialogConfig {
-    data class ProductCard(val productId: Int) : DialogConfig()
+    data class ProductCard(val productId: Long) : DialogConfig()
+
     data object DeleteUser : DialogConfig()
+
     data object LogoutUser : DialogConfig()
 }
